+++ src/config/getBabelOptions.js
@@ -22,16 +22,11 @@ module.exports = function(
       isES6Enabled && [
         require("@babel/preset-env").default,
         {
-          // targets: {
-          //   ie: 9
-          // },
           ignoreBrowserslistConfig: true,
           // useBuiltIns: "entry",
           useBuiltIns: false,
           targets: { browsers: ["Android >= 4.0", "ios >= 8", "ie >=9"] },
           modules: isModule ? false : "amd"
-          // modules: false,
-          // modules: 'amd'
           // debug: true
         }
       ],
@@ -45,10 +40,14 @@ module.exports = function(
     ].filter(Boolean),
     plugins: [
       [require("../babel-plugin-tinytool")],
-      // isES6Enabled && [require('babel-plugin-transform-class-properties')],
-      // isES6Enabled && [require('babel-plugin-transform-object-rest-spread')],
-      // isES6Enabled && [require('babel-plugin-transform-export-extensions')],
-      // isES6Enabled && [require('babel-plugin-transform-decorators-legacy').default],
+      isES6Enabled && [
+        require("@babel/plugin-proposal-decorators").default,
+        { legacy: true }
+      ],
+      isES6Enabled && [
+        require("@babel/plugin-proposal-class-properties").default,
+        { legacy: true }
+      ],
       isES6Enabled && [
         require("@babel/plugin-transform-runtime").default,
         {